refactor(child-pi): H-7 step 4 — extract stdout/stderr stream parsing to child-pi-streams.ts
Decomposed the ChildPiLineObserver class and its 5 helper functions
(compactContentPart, compactChildPiEvent, displayTextFromCompactEvent,
nonJsonLineResult, compactChildPiLine) into a dedicated module. The code
is a verbatim copy of what lived in child-pi.ts — only imports were updated
to pull compactString/compactValue/appendTranscript/flushPendingTranscriptWrites
from child-pi-transcript.ts and MAX_*_CHARS/MAX_LINE_BUFFER_BYTES from
child-pi-constants.ts.

child-pi.ts: 1555 → 1279 lines (-276). New child-pi-streams.ts: 277 lines.
Public API (ChildPiLineObserver) preserved via re-export from child-pi.ts.

All 91 child-pi + lock + dependent tests pass (no regressions).
Verified baseline + modified: identical behavior on all test cases.
